requirements.txt
langchain-google-genai==2.1.0
langchain==0.3.21
python-dotenv==1.1.0
pandas==2.2.3
langchain-experimental==0.3.4
tabulate==0.9.0
BeautifulSoup==4.13.0
langchain_google_genai==1.3.1
langchain-core==0.3.50
langchain.tools==0.1.34
